Refactor rendering into modular components
